sending battery voltage in telemetry

+++ include/io.h
@@ -24,7 +24,7 @@ void io_pwr_init(void);
 void io_ext_watchdog_config(void);
 void io_ext_watchdog_service(void);
 
-void io_vbat_meas_init(int8_t a_coeff, int8_t b_coeff);
+void io_vbat_meas_init(int16_t a_coeff, int16_t b_coeff);
 uint16_t io_vbat_meas_get(void);
